Rename of main file, remove runs specifier
This feature has been removed to avoid optimizations in the external
for loop, used for re-running the same code more than once.
Instead, to accomplish the same theoretical behaviour, just run the
code more times.
The main file has been renamed to get a better name fit with
the code inside it.