Add implementation for _Imp1 (prefetching)
Moved the save of result CSV into a function in "utility_functions.c".
Implemented code with prefetch(not very effective).
It worsen the performance, probably because it add additionals
instruction to do the prefetching, but the hardware prefetcher
is good enough and so these are only more instructions to run.
Added printing of function name when DEBUG is set >= 1.