Add DCT (Differential Comparison Test)
A DCT is a type of test that passes when a program/function/etc returns
the same results of the reference one, otherwise it will fail.
Added a macro that override the DEBUG macro definition, and disables
the possibility to output results.
The only output on stdout will be if the current implementation gives
the same results of the golden one selected.